Add launch file for rll path planning for kuka
Actually launched the project and was able to see kuka setup on both gazebo and rviz!
Rviz also has the basic (default) code running so the arm moves down a bit towards the maze (in rviz)

There are some errors in the gazebo window, where it complains that a joint controller interface is not loaded.
I have captured these errors in logs/error.1.text