Add an optional project messages / error logger
Interestingly some of the spewed output is binary but displays
well on console.

With resorting path planner to default (or not), basic grasping
still fails (trajectory path takes too long to execute) although
visually the path to pick the object seems to work. It may be that
something is just blocking the grasping action part.